+++ CHANGES.md
@@ -47,6 +47,8 @@ To be released.
  -  Changed `TxSuccess.UpdatedStates`'s type to
     `IImmutableDictionary<Address, IValue>` from
     `IImmutableDictionary<Address, IValue?>`.  [[#3248]]
+ -  Added `IBlockChainStates.GetState()` interface method.  [[#3250]]
+ -  Added `IBlockStates.GetState()` interface method.  [[#3250]]
 
 ### Backward-incompatible network protocol changes
 
@@ -70,6 +72,7 @@ To be released.
 [#3245]: https://github.com/planetarium/libplanet/pull/3245
 [#3247]: https://github.com/planetarium/libplanet/pull/3247
 [#3248]: https://github.com/planetarium/libplanet/pull/3248
+[#3250]: https://github.com/planetarium/libplanet/pull/3250
 
 
 Version 2.2.0